增加了邮件发送功能

+++ code/BCJ/user/views.py
@@ -7,7 +7,7 @@ import json
 import uuid
 import random
 import datetime
-
+import pytz
 from . import models
 # Create your views here.
 def register(request):
@@ -68,16 +68,17 @@ def findPassword(request):
     data = json.loads(request.body)
     Email = data['Email']
     user = models.User.objects.filter(Email=data['Email']).first()
-
-    email_title = "找回密码"
-    code = random_str()             #随机生成的验证码
-    IdentifyingCode = models.IdentifyingCode.objects.create(Code=code)
-    IdentifyingCode.User.add(user)
-    email_body = "验证码为: {0}".format(code)
-    t1 = Thread( target = send_mail, args = (email_body, settings.EMAIL_HOST_USER,[Email]) ) 
-    t1.start()
-
-    return HttpResponse("验证码已发送，请查收邮件")
+    if user:
+        email_title = "找回密码"
+        code = random_str()             #随机生成的验证码
+        IdentifyingCode = models.IdentifyingCode.objects.create(Code=code)
+        IdentifyingCode.User.add(user)
+        email_body = "验证码为: {0}".format(code)
+        t1 = Thread( target = send_mail, args = (email_title, email_body, settings.EMAIL_HOST_USER,[Email]) ) 
+        t1.start()
+        return HttpResponse("验证码已发送，请查收邮件")
+    else:
+        return HttpResponse("邮箱未注册")
 
 
 def Verifycode(request):
@@ -85,18 +86,29 @@ def Verifycode(request):
     Email = data['Email']
     Newpassword =data['Newpassword']
     user = models.User.objects.filter(Email=data['Email']).first()
-    user_code = list(models.IdentifyingCode.objects.filter(User=user))
-    if user_code:
-        now_time = datetime.datetime.now()
-        code = data['Code']                     #获取传递过来的验证码
-        for item in user_code:
-            if (now_time-item.Time).seconds > 120 and code == item.Code:
-                user.Password = Newpassword
-                msg = "密码已重置"
-            item.delete()
-        if msg != "密码已重置":
-            msg = "验证码已过期，请重新获取"
-        return HttpResponse(msg)   
+    if user:
+        user_code = list(models.IdentifyingCode.objects.filter(User=user))
+        if user_code:
+            now_time = datetime.datetime.now(pytz.utc)
+            code = data['Checksum']                     #获取传递过来的验证码
+            msg = ""
+            for item in user_code:
+                diff = (now_time-item.Time).seconds
+                if 0 <= diff and diff < 1800 :
+                    if code == item.Code:
+                        user.Password = Newpassword
+                        user.save()
+                        msg = "设置成功"
+                        item.delete()
+                        print("设置成功")
+                else:
+                    item.delete()
+            if msg == "":
+                msg = "验证码错误"
+            
+            return HttpResponse(msg)   
+        else:
+            return HttpResponse("邮箱无验证码")
     else:
-        return HttpResponse("邮箱错误")
+        return HttpResponse("邮箱未注册")
